Set up prettier
Prettier is a code formatter. It takes code as input and spits out the same code with a prettier format.

It drops most the original styling and formats the code from its syntax tree, the output style is always consistent.

The `test:lint` task asserts that all of the `.js` files match both the current eslint rules and are formatted correctly according to Prettier. It is necessary to run Prettier to reformat any new changes in-place before committing.

**Using Prettier**

Run `npm run format` before committing to reformat any changed code, _or_ configure your editor to automatically run Prettier as files are changed:

**Sublime Text 3:**
  - Install [JsPrettier](https://github.com/jonlabelle/SublimeJsPrettier) (there is not need to install `prettier` globally, the local dependency will be used)
  - Configure JsPrettier to auto format files on save:
  ```json
    // ~/Library/Application Support/Sublime Text 3/Packages/User/JsPrettier.sublime-settings

    {
      "auto_format_on_save": true,
      "auto_format_on_save_excludes": [
        "*/package.json"
      ]
    }
  ```

**Atom:**
 - Install [Prettier for Atom](https://atom.io/packages/prettier-atom)
 - Enable "format on save" (Packages → Prettier → Toggle Format on Save)